# blockchain/docs/blockchain/swagger.yaml
basePath: /
definitions:
  dto.SendRequest:
    properties:
      amount:
        type: number
      to:
        type: string
    required:
    - amount
    - to
    type: object
  dto.TopupRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service that handles all blockchain request
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blockchain service
  version: "1.0"
paths:
  /v1/blockchain/wallet:
    get:
      consumes:
      - application/json
      description: Retrieve a wallet from the blockchain for a specific user
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wallet address
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a wallet by user ID
      tags:
      - Blockchain
  /v1/blockchain/wallet/balance:
    get:
      consumes:
      - application/json
      description: Retrieve the balance of a specific address on the blockchain
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balance
          schema:
            type: number
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get the balance of an address
      tags:
      - Blockchain
  /v1/blockchain/wallet/balance/address:
    get:
      consumes:
      - application/json
      description: Retrieve the balance of a specific address on the blockchain
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet address
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balance
          schema:
            type: number
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get the balance of an address
      tags:
      - Blockchain
  /v1/blockchain/wallet/create:
    post:
      consumes:
      - application/json
      description: Create a new wallet on the blockchain
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: New Wallet
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new wallet
      tags:
      - Blockchain
  /v1/blockchain/wallet/qr:
    get:
      consumes:
      - application/json
      description: Return a wallet QR code by user ID
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wallet QR code
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a wallet QR code by user ID
      tags:
      - Blockchain
  /v1/blockchain/wallet/transactions:
    post:
      consumes:
      - application/json
      description: Send cryptocurrency from one address to another on the blockchain
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Send Request
        in: body
        name: sendRequest
        required: true
        schema:
          $ref: '#/definitions/dto.SendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Send cryptocurrency to another address
      tags:
      - Blockchain
    put:
      consumes:
      - application/json
      description: TopUp top up of an account
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Top up Request
        in: body
        name: topUpRequest
        required: true
        schema:
          $ref: '#/definitions/dto.TopupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: TopUp top up of an account
      tags:
      - Blockchain
  /v1/blockchain/wallet/usd/balance:
    get:
      consumes:
      - application/json
      description: Retrieve the balance in USD of a specific address on the blockchain
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balance in USD
          schema:
            type: number
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get the balance in USD of an address
      tags:
      - Blockchain
schemes:
- http.
swagger: "2.0"